requirements.txt
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
blinker==1.9.0
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1734473278428/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
conda @ file:///croot/conda_1734648259968/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1733258728466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1732129644106/work
cycler==0.12.1
distro @ file:///croot/distro_1714488253808/work
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
frozendict @ file:///croot/frozendict_1713194832637/work
fsspec==2025.2.0
idna @ file:///croot/idna_1714398848350/work
itsdangerous==2.2.0
Jinja2==3.1.5
joblib @ file:///croot/joblib_1718217211762/work
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
kiwisolver==1.4.8
libmambapy @ file:///croot/mamba-split_1732896306512/work/libmambapy
MarkupSafe==3.0.2
matplotlib==3.10.0
menuinst @ file:///croot/menuinst_1731364904452/work
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
networkx==3.4.2
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp312-cp312-linux_x86_64.whl#sha256=e2374991344fa2241a2153ef3d550d3a1cd2d50cb1f1d51eddc82870abc88021
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyparsing==3.2.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
PyYAML @ file:///croot/pyyaml_1728657952215/work
requests @ file:///croot/requests_1730999120400/work
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
scikit-learn==1.6.0
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
torch==2.6.0
torchaudio==2.5.1
torchvision==0.20.1
tqdm @ file:///croot/tqdm_1724853939799/work
triton==3.2.0
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1727769808118/work
Werkzeug==3.1.3
wheel==0.44.0
zstandard @ file:///croot/zstandard_1731356346222/work
